requirements.txt
numpy>=1.23.5
pillow~=10.4.0